fix(review): isolate specialist discovery
